appengine/chromium_status/requirements.txt
requests==2.25.1
googleapis_common_protos
google-cloud-ndb